backend/requirements/base.txt
asgiref==3.2.10
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.1.8
django-allauth==0.44.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.12.1
drf-spectacular==0.15.1
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==5.6.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
oauthlib==3.1.0
packaging==21.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
responses==0.20.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
xmltodict==0.12.0
